[datafeeder] Fix CRS handling
GeoServer published CRS were wrong sometimes.

Make sure that the following procedd is followed:

When publishing an uploaded shapefile:

- If it has no .prj file, the publish request MUST provide the SRS code
- If the .prj matches a known EPSG code:
  - if the publish request provides an SRS, the request one is used,
    both when importing the data to PostGIS and as the GS layer CRS.
  - if the publish request does not provide an SRS, the recognized
    native one from the .prj is used both to import and for the GS layer.
- If the .prj does NOT match a known EPSG code:
  - the publish request MUST provide an SRS, and that one is used for
    the imported dataset (PostGIS) and as the GS layer CRS.

Note: reprojection from the dataset native CRS to the one requested in
the publish request is NOT implemented, though can be easily be done.
But it's not a requirement right now, despite most of the code being
there, except the reprojection itself in DatasetsService.